fix(ci,pkg-web,b1,b4): rebuild stale wasm + wire FuelTrancheRunner + selftest
- pkg-web wasm was stale (Jul15) and missing kernel exports: root cause was
  lib.rs hand-listed 'pub use wasm::{5 items}' (dropped 19 of 24 wasm-bindgen
  fns) + hydra.rs used usize::from_le_bytes on a [u8;8] (broke wasm32's
  32-bit usize). Fixed: pub use wasm::* (glob) + u64::from_le_bytes(g) as
  usize; clean-rebuilt kernel --target wasm32 --features wasm and regenerated
  pkg-web via wasm-bindgen. All 24 wasm exports now present (was 22, 0 _js
  in the broken empty wasm).
- B1: wire FuelTrancheRunner into AgentDispatcher::dispatch execution path
  (was a dead module). Uses DeterministicFuelMeter with 1-unit tranche for
  exact-cost budget envelope, reusing the production fuel-control loop. 19+1
  agent-adapters tests green.
- B4: add criterion fuel-bench (fuel_bench.rs) calibrating the FUEL_PER_UNIT
  placeholder; documents that real calibration needs a wasmtime-fuel guest.
- CI: add tools/telemetry/selftest-telemetry.sh (was missing -> ci.yml
  telemetry-selftest job failed on a nonexistent file). Syntax-checks all
  scripts; skips live send test when the untracked binary is absent.

Verified: kernel native 485 tests green; agent-adapters 19+1 + bench compiles;
engine builds; pkg-web 24/24 exports; selftest script OK.

+++ kernel/src/hydra.rs
@@ -114,7 +114,7 @@ impl BreachAlert {
         g.copy_from_slice(&b[32..]);
         Some(BreachAlert {
             node_id,
-            group_size: usize::from_le_bytes(g),
+            group_size: u64::from_le_bytes(g) as usize,
         })
     }
 }
